Add the admin panel
Admin panel also includes a websocket channel that monitors the
changes in the mongodb imboard_info collection and updates the panel
accordingly. You can start or stop indexing from the admin panel. In
the future it will also be the place to start or stop image
downloading and configure more things about indexing