Implement scrapper configuration file
The files lives in the root directory and is named scrappers.toml, it
contains scrapper definitions in a simple format. The file also
contains some documentation about how the scrappers are defined.

All the classes that represented scrapper types are now made
concrete/non-abstract, as they are now used to spawn scrappers from
the configuration file insted of being inherited by other classes to
make scrappers in code. This also means that some parameters are now
provided as constructor arguments instead of fields that are overriden
by children classes.

The toml library snaphot is local and is used  because the
library contains a bug that was fixed locally, but is not yet
merged upstream