Implement searching by image properties
The syntax is "propety >(=)/<(=)/= value". Whitespaces are not
significant, the syntax is always STRING OPERATOR STRING/INT and after
that other tags are parsed again.

Tags are not encoded on the client side when searching any more,
because that breaks > and < operators, instead they are now encoded
on the server side where appropriate.

Image now has a companion object with an intField function to help
find image properties that can use >/< operators, using those
operators on some other property will throw an exception.

>/</= are now prohibited in simple tags.